+++ CHANGELOG.md
@@ -15,3 +15,15 @@ often, sometimes several times in a day. This also means that any
 individual release will not have many changes within it. Below is a list
 of the releases we've made so far, along with what was changed within
 each release.
+
+## 0.3.0 (2026-03-25)
+
+### Feat
+
+- :sparkles: add `publish` (#32)
+
+## 0.2.0 (2026-03-23)
+
+### Feat
+
+- :sparkles: move existing code from old repo (#28)